requirements.txt
streamlit==1.11.0
altair<5
rdkit
requests
STOUT-pypi